# openapi.yaml
openapi: 3.0.3
info:
  title: WarrantyVault API
  description: |
    REST API for native mobile clients (iOS / Android) of WarrantyVault.
    Source-of-truth for client codegen:
      - iOS:     swift-openapi-generator
      - Android: openapi-generator-cli (kotlin)
    The web app continues to use Next.js Server Actions — those are NOT
    documented here.

    Auth: bearer token from POST /api/v1/auth/login (or /register). Send as
    `Authorization: Bearer <token>` on every authenticated request.

    Ngôn ngữ: mọi chữ trong `message` / `fieldErrors` được server dịch sẵn. Thứ
    tự quyết định: `?lang=vi|en` → header `Accept-Language` → `User.locale` (đã
    lưu) → `en`. Ngôn ngữ không hỗ trợ rơi về mặc định, **không** phải lỗi 400.
    Xem `components.parameters.Lang`.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://warrantyvault.example.com
    description: Production (set when deployed)
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/push/register:
    post:
      tags: [push]
      summary: Register a push token (web / apns / fcm)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/WebPushInput'
                - $ref: '#/components/schemas/NativePushInput'
      responses:
        '201': { $ref: '#/components/responses/Ok' }
        '400': { $ref: '#/components/responses/BadInput' }
...
  /api/v1/push/{id}:
    parameters: [{ $ref: '#/components/parameters/IdPath' }]
    delete:
      tags: [push]
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string }

    # The language of the `message` / `fieldErrors` COPY in a response. It is not
    # a resource: it changes no status code, no error code and no field name, so a
    # client that ignores it loses nothing but the localisation.
    #
    # Precedence, first match wins:
    #   1. this parameter (`vi` | `en`)
    #   2. the `Accept-Language` request header (q-weighted, parsed by
    #      golang.org/x/text/language)
    #   3. the caller's stored preference (`User.locale`, set through PATCH
    #      /api/v1/auth/me)
    #   4. `en`
    #
    # A language we do not ship is a FALLBACK, never an error: `?lang=fr` is
    # ignored (the next level decides) and `Accept-Language: fr` gets English with
    # a 200. The response carries `Content-Language` when the request asked for a
    # language at all, so a client can see what it actually got.
...
  responses:
    Ok:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties: { ok: { type: boolean } }
...
    BadInput:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    NotFound:
      description: Resource not found (or owned by another user)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string, description: 'Stable code, e.g. unauthorized, not_found, bad_input, rate_limited, limit_reached' }
        message: { type: string, description: 'Human-readable Vietnamese message' }
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
...
    WebPushInput:
      type: object
      required: [endpoint, p256dh, auth]
      properties:
        platform: { type: string, enum: [web], default: 'web' }
        endpoint: { type: string, format: uri }
        p256dh: { type: string }
        auth: { type: string }
        userAgent: { type: string, maxLength: 500, nullable: true }
    NativePushInput:
      type: object
      required: [platform, token]
      properties:
        platform: { type: string, enum: [apns, fcm] }
        token: { type: string, minLength: 8, maxLength: 4096 }
        userAgent: { type: string, maxLength: 500, nullable: true }
...